refactor(web): split task switcher components
