+++ apps/web/components/task/task-switcher-types.ts
@@ -1,0 +1,93 @@
+import type { ForegroundActivity, TaskSessionState, TaskState } from "@/lib/types/http";
+import type { GroupedSidebarList } from "@/lib/sidebar/apply-view";
+import type { TaskMoveWorkflow } from "@/components/task/task-move-context-menu";
+
+export type StepDef = {
+  id: string;
+  title: string;
+  color?: string;
+  events?: { on_enter?: Array<{ type: string; config?: Record<string, unknown> }> };
+};
+
+export type TaskLinkHandler = (taskId: string, taskTitle?: string) => void;
+
+export type TaskSwitcherItem = {
+  id: string;
+  title: string;
+  state?: TaskState;
+  sessionState?: TaskSessionState;
+  /** Task-level most-active-wins busy aggregate (ADR-0049) from the task record. */
+  foregroundActivity?: ForegroundActivity | null;
+  description?: string;
+  workflowId?: string;
+  workflowName?: string;
+  workflowStepId?: string;
+  workflowStepTitle?: string;
+  repositoryPath?: string;
+  repositories?: string[];
+  diffStats?: { additions: number; deletions: number };
+  isRemoteExecutor?: boolean;
+  remoteExecutorType?: string;
+  remoteExecutorName?: string;
+  updatedAt?: string;
+  createdAt?: string;
+  isArchived?: boolean;
+  primarySessionId?: string | null;
+  hasPendingClarification?: boolean;
+  hasPendingPermission?: boolean;
+  parentTaskTitle?: string;
+  parentTaskId?: string;
+  workspaceMode?: "inherit_parent" | "new_workspace" | "shared_group";
+  prInfo?: { number: number; state: string; aggregateState?: string };
+  isPRReview?: boolean;
+  isIssueWatch?: boolean;
+  issueInfo?: { url: string; number: number };
+  agentErrorMessage?: string | null;
+};
+
+export type TaskSwitcherProps = {
+  grouped: GroupedSidebarList;
+  workflows?: TaskMoveWorkflow[];
+  stepsByWorkflowId?: Record<string, StepDef[]>;
+  activeTaskId: string | null;
+  selectedTaskId: string | null;
+  collapsedGroupKeys?: string[];
+  onToggleGroup?: (groupKey: string) => void;
+  collapsedSubtaskParentIds?: string[];
+  onToggleSubtasks?: (parentTaskId: string) => void;
+  onSelectTask: (taskId: string) => void;
+  onEditTask?: (task: TaskSwitcherItem) => void;
+  onRenameTask?: (taskId: string, currentTitle: string) => void;
+  onArchiveTask?: (taskId: string) => void;
+  onCreateSubtask?: (taskId: string, taskTitle: string) => void;
+  onDeleteTask?: (taskId: string) => void;
+  onDetachTask?: (taskId: string) => void;
+  onLinkPullRequest?: TaskLinkHandler;
+  onLinkIssue?: TaskLinkHandler;
+  onLinkMergeRequest?: TaskLinkHandler;
+  onLinkJiraTicket?: TaskLinkHandler;
+  onLinkLinearIssue?: TaskLinkHandler;
+  onLinkSentryIssue?: TaskLinkHandler;
+  onMoveToStep?: (taskId: string, workflowId: string, targetStepId: string) => void;
+  onTogglePin?: (taskId: string) => void;
+  onReorderGroup?: (groupTaskIds: string[]) => void;
+  onReorderSubtasks?: (parentTaskId: string, orderedSubtaskIds: string[]) => void;
+  pinnedTaskIds?: string[];
+  deletingTaskId?: string | null;
+  isLoading?: boolean;
+  loadError?: string | null;
+  onRetryLoad?: () => void;
+  retryLabel?: string;
+  totalTaskCount?: number;
+  // Multi-select (cmd/shift click). When the selection is non-empty, plain
+  // clicks toggle instead of navigating; the context menu acts on the selection.
+  selectedTaskIds?: Set<string>;
+  onToggleSelectTask?: (taskId: string) => void;
+  onSelectTaskRange?: (taskId: string) => void;
+  onBulkArchive?: (taskIds: string[]) => void;
+  onBulkDelete?: (taskIds: string[]) => void;
+  onBulkPin?: (taskIds: string[]) => void;
+  onBulkMove?: (taskIds: string[], targetWorkflowId: string, targetStepId: string) => void;
+  onClearSelection?: () => void;
+  isMixedWorkflowSelection?: boolean;
+};
